requirements.txt
psycopg2-binary==2.9.10
google-genai==1.30.0
fastmcp==2.11.3
